[consensus] Move IsFinalTx from tx_verify.h to tx_check.h
Context-independent IsFinalTx doesn't rely on chain or mempool state
so moving it to tx_check let it being used by wallet code without
linking server code in following commits.
