+++ src/miner.cpp
@@ -11,6 +11,7 @@
 #include <coins.h>
 #include <consensus/consensus.h>
 #include <consensus/merkle.h>
+#include <consensus/tx_check.h>
 #include <consensus/tx_verify.h>
 #include <consensus/validation.h>
 #include <policy/feerate.h>
