+++ src/test/miner_tests.cpp
@@ -6,6 +6,7 @@
 #include <coins.h>
 #include <consensus/consensus.h>
 #include <consensus/merkle.h>
+#include <consensus/tx_check.h>
 #include <consensus/tx_verify.h>
 #include <miner.h>
 #include <policy/policy.h>
